Add Chromium vulnerabilities
The list of vulnerabilities was obtained by searching for
'cpe:/a:google:chrome' in the 'vuln:vulnerable-software-list' element of
the NVD data feed files. The bounty information was identified using the
data collected for the CVSS versus Bounty Research project. Finally, the
date of publication of a blog post was identified using the Google
Blogger API.