Report mistakes made regarding the vulnerability
I decided to discuss what caused the vulnerability to persist rather
than go in-depth with the mistake creating the vulnerability itself. I
believe this is a more potent way to go about it, since the
vulnerability fix was literally one line long and the story of the
persistence of the vulnerability may prove more interesting to the
software engineering community.