special handling of "mega" commits
So this was designed on the fly, but it SIGNIFICANTLY reduces the overhead of over 300k files from the initial import of the project. I looked at commits that had over 1,000 files in their diff and made them into a "mega commit" file. This has a special curator note. It'll still show up on the timeline, it just wont' have the filepath information (which is probably fine anyway).

This should (a) increase loading time, (b) improve database peformance overall